Updated scan table for au-Sydney_North_Shore
Hi,

Please find attached an updated scan table for au-Sydney_North_Shore. The
main change is that the frequency for channel SBS is now 184500000[1].

The comment for channel D44 has been changed to "TVS UHF29". D44 is the
name of the trial data casting service that ended in 2010[2]. The community
channel TVS now broadcasts on this service[3].

[1] http://www20.sbs.com.au/transmissions/index.php?pid=2&id=902
[2] http://en.wikipedia.org/wiki/Digital_Forty_Four
[3] http://www.tvs.org.au/info/tune-in

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>