Change Sydney FEC settings to AUTO.
The FEC settings are not stipulated by the government as part of the
license and a number of broadcasters claim that the transmission
parameters are subject to change.

Additionally, the old settings were fec_lo=NONE, which is not completely
accurate, since the observed values are never NONE. The code in scan
will substitute FEC_AUTO for FEC_NONE, but we might as well be explicit.