Remove zeros on the left for channel frequencies
Those zeros are there just to align the entries. However, we'll
convert them to DVBv5, and the conversion tools won't keep the
same alignment.

This is not needed for the conversion itself, but we want to
do the conversion on both directions and see if the file will
remain the same. Stripping those useless zeros makes easier
for us to double check if something got badly converted.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>