renaming NTSC file because it doesnt actually use 8VSV, updating channel listings, add channel numbers
This file is renamed to more accurately reflect its contents, as well as including all channels every used in NTSC. This is primarily of historical interest, the ATSC file has a more up-to-date (and much shorter) list to scan against. I notice that my local repo has the rm, but this isn’t reflected in the patch (of course), is there something else I should have done as well?

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>