dk-YouSee
I've compiled a list of channel/transponders as used by the big Danish
cable provider Yousee, for use with tvheadend, a tv server and recorder.
I asked the people behind the program to put the list in the program,
but they told me that they use your repository, and that I therefore
should ask you. Is there any way to add this file to your list,
dtv-scan-tables.git, at linuxtv.org?