Errors in scan file
I posted an email with a patch back on 04/02/2016 17:32 only to find that nothing has been done.

This patch, reproduced below adds 2 new transponders to Astra 28.2E and removes a trailing space. This trailing spaces causes dvbv5-scan to segfault.

Why has this not been actioned and why was my email address banned? Is this a close community that tells people to f-off?

dvb-s/Astra-28.2E | 27 ++++++++++++++++++++++++---

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>